Install Eventing by default and scale it down when unused
POC to install a default KnativeEventing instance
that scales down components when they are not used just to prove that
the idea of installing a component by default and scale it down when
it's not used is doable.
  It's not exactly scale down to 0 because webhooks need to stay up in
order to be able to create resources. The scaling is pretty quick,
I've tried with a PingSource and getting your first PingSource to
Ready takes a few seconds (<10s) but the installation steps are
reduced to just installing the operator. The memory usage for running
idle webhooks is <200MB and for CPU < 0.01.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>